moved interfaces and types to file with types
