+++ src/components/AgenciesPage/actions/ActionTypesAgencies.ts
@@ -1,4 +1,4 @@
-export const enum AgenciesActions {
+export const enum AgenciesActionTypes {
   FETCH_AGENCIES_REQUEST = 'FETCH_AGENCIES_REQUEST',
   FETCH_AGENCIES_SUCCESS = 'FETCH_AGENCIES_SUCCESS',
   FETCH_AGENCIES_FAILURE = 'FETCH_AGENCIES_FAILURE ',
