+++ src/components/AgenciesPage/types/Agencies.ts
@@ -1,3 +1,5 @@
+import { AgenciesActionTypes } from '../actions/ActionTypesAgencies';
+
 export interface Agency {
   id: number;
   name: string;
@@ -12,3 +14,18 @@ export interface Agency {
 
 export type AgencyItemContent = Omit<Agency, 'id' | 'image_url'>;
 export type AgenciesSpecification = Omit<Agency, 'id' | 'image_url' | 'name'>;
+
+export interface AgenciesState {
+  agencies: Agency[];
+  isLoading: boolean;
+  error: string;
+}
+
+export interface AgenciesAction {
+  type: AgenciesActionTypes;
+  payload?: AgenciesState[keyof AgenciesState];
+}
+
+export type ActionCreatorRequest = () => AgenciesAction;
+export type ActionCreatorSuccess = (agencies: Agency[]) => AgenciesAction;
+export type ActionCreatorFailed = (error: string) => AgenciesAction;
